requirements.txt
jinja2
pyyaml
python-box